fix landing page
