+++ src/@/entities/models/entity.model.ts
@@ -26,6 +26,19 @@ export const mergeEntityRegistryFromApi = createEvent<Partial<Record<string, Par
 
 export const $entityRegistry = createStore<Record<string, EntityConfig>>({ ...DEFAULT_ENTITY_REGISTRY });
 
+/** Registry filtered to only active entities (from config.active flag), in registry order */
+export const $entityRegistryFiltered = $entityRegistry.map((registry) => {
+  const entries = Object.entries(registry).filter(([, config]) => config.visible);
+  return Object.fromEntries(entries) as Record<EntityType, EntityConfig>;
+});
+
+/** Active entity types in registry order (based on config.active flag) */
+export const $entityTypesFiltered = $entityRegistry.map((registry) => {
+  return Object.entries(registry)
+    .filter(([, config]) => config.visible)
+    .map(([type]) => type as EntityType);
+});
+
 // Full replacement (e.g., reset)
 $entityRegistry.on(updateEntityRegistry, (_, payload) => payload);
 
@@ -73,13 +86,14 @@ export const $entityConfigMap = $entityRegistry;
  */
 export const changeActiveEntityTypes = createEvent<EntityType[]>();
 export const setActiveEntityTypes = changeActiveEntityTypes; // alias for route model sync
-export const $activeEntityTypes = createStore<EntityType[]>(['school']);
+export const $activeEntityTypes = createStore<EntityType[]>(['health', 'school']);
 $activeEntityTypes.on(changeActiveEntityTypes, setPayload);
 
 /**
  * Toggle a single entity type on/off in the active list.
  */
 export const toggleEntityType = createEvent<EntityType>();
+
 $activeEntityTypes.on(toggleEntityType, (current, entityType) => {
   if (current.includes(entityType)) {
     if (current.length <= 1) return current;
@@ -157,3 +171,4 @@ export const $activeEntityConfigs = combine(
     .map(type => registry[type])
     .filter((config): config is EntityConfig => Boolean(config))
 );
+
