+++ src/core/routes.ts
@@ -20,7 +20,7 @@ export const mapOverview = router.add('/map');
 export const mapCountries = router.add('/map/countries:path(/.*)?');
 export const mapCountry = router.add<Code>('/map/country/:code:path(/.*)?');
 export const mapSchools = router.add('/map/schools:path(/.*)?');
-export const mapEntities = router.add('/map/entities:path(/.*)?');
+export const entityView = router.add('/map/view:path(/.*)?');
 export const aboutus = router.add('/about');
 export const tour = router.add('/map/tour');
 
@@ -114,7 +114,7 @@ export const $mapRoutes = combine({
   map: mapOverview.visible,
   country: mapCountry.visible,
   schools: mapSchools.visible,
-  entities: mapEntities.visible,
+  entityView: entityView.visible,
 });
 
 export const $countryRouteLeave = guard(mapCountry.visible, { filter: (value) => !value })
